Refactor content filters
- Generalize page filters to all content types with translations
- Replace filters in templates
