+++ backend/cms/templates/pages/page.html
@@ -3,6 +3,7 @@
 {% block content %}
 {% load static %}
 {% load widget_tweaks %}
+{% load content_filters %}
 {% load page_filters %}
 {% load rules %}
 {% load compress %}
@@ -43,7 +44,7 @@
                         {% get_current_language as LANGUAGE_CODE %}
                         {% unify_language_code LANGUAGE_CODE as LANGUAGE_CODE %}
                         {% if LANGUAGE_CODE != language.code %}
-                            {% get_page_translation page LANGUAGE_CODE as backend_translation %}
+                            {% get_translation page LANGUAGE_CODE as backend_translation %}
                             {% if backend_translation %}
                                 ({% trans 'Title in' %} {% translated_language_name LANGUAGE_CODE %}: "{{ backend_translation.title }}")
                             {% endif %}
