+++ backend/cms/templates/pages/tree_row.html
@@ -1,4 +1,5 @@
 {% load i18n %}
+{% load content_filters %}
 {% load page_filters %}
 <tr data-drop-id="{{ node.id }}" data-drop-position="left" class="drop drop-between h-3 -m-3 hidden level-{{page.depth}}"><td colspan="9"><div><span></span></div></td></tr>
 <tr data-drop-id="{{ node.id }}" data-drop-position="last-child" class="drop drop-on border-t border-b border-solid border-grey-lighter hover:bg-grey-lightest{% if page.depth > 0 %} child level-{{page.depth}}{% endif %}">
@@ -34,7 +35,7 @@
     {% get_current_language as LANGUAGE_CODE %}
     {% unify_language_code LANGUAGE_CODE as LANGUAGE_CODE %}
     {% if LANGUAGE_CODE != language.code %}
-        {% get_page_translation page LANGUAGE_CODE as backend_translation %}
+        {% get_translation page LANGUAGE_CODE as backend_translation %}
         <td>
             <a href="{% url 'edit_page' page_id=page.id region_slug=region.slug language_code=LANGUAGE_CODE %}" class="block py-3 px-2 text-grey-darkest">
                 {% if backend_translation %}
@@ -50,7 +51,7 @@
             <div class="lang-grid">
                 {% for other_language in languages %}
                     <a href="{% url 'edit_page' page_id=page.id region_slug=region.slug language_code=other_language.code %}">
-                        {% get_page_translation page other_language.code as other_translation %}
+                        {% get_translation page other_language.code as other_translation %}
                         {% if other_translation %}
                             {% if other_translation.currently_in_translation %}
                                 <span title="{% trans 'Currently in translation' %}">
